pyforge-django/pyproject.toml
# Author: Abdulwahed Mansour
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "pyforge-django"
version = "0.1.0"
description = "Rust-accelerated Django serialization, validation, and field mapping"
authors = [{ name = "Abdulwahed Mansour" }]
license = { text = "MIT OR Apache-2.0" }
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Rust",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django>=4.2",
]

[project.optional-dependencies]
drf = ["djangorestframework>=3.14"]
dev = ["pytest>=7.0", "pytest-django>=4.5"]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "."
module-name = "pyforge_django"
